requirements.txt
psutil
numba
numpy
pyautogui
opencv-python
pillow
pygetwindow
pytesseract
matplotlib